remove: setup_with_defaults()
The original idea was for a developer to use setup_with_defaults() before initializing the actual logger. This way, there was a default logger implementation, and the developer could init their logger after loading user config files or similar. But a global logger in rust can only mit set once. Instead, this crate provides macros to log, which will fall back to using stdout/stderr when logging and no logger has been initialized yet. This way, the logging macros of this crate will always produce output at the cost of a bit of runtime performance.